User-manager-API/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
certifi==2023.11.17; python_version >= '3.6'
cffi==1.16.0; python_version >= '3.8'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cryptography==41.0.7; python_version >= '3.7'
defusedxml==0.8.0rc2; python_version >= '3.6'
django==5.0.1; python_version >= '3.10'
django-templated-mail==1.1.1
djangorestframework==3.14.0; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
djoser==2.2.2; python_version >= '3.8' and python_version < '4.0'
idna==3.6; python_version >= '3.5'
oauthlib==3.2.2; python_version >= '3.6'
pycparser==2.21
pyjwt==2.8.0; python_version >= '3.7'
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0; python_version >= '3.7'
requests-oauthlib==1.3.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
social-auth-app-django==5.4.0; python_version >= '3.8'
social-auth-core==4.5.1; python_version >= '3.8'
sqlparse==0.4.4; python_version >= '3.5'
tzdata==2023.4; sys_platform == 'win32'
urllib3==2.1.0; python_version >= '3.8'
